requirements.txt
# Core Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Genomics and VCF Processing
pysam==0.22.0
cyvcf2==0.30.18
biopython==1.81

# Database and Storage
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP and API
httpx==0.25.2
aiofiles==23.2.1

# Data Processing and Analysis
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Production Dependencies
gunicorn==21.2.0
psycopg2-binary==2.9.9
